# fix sql-server dbm error in certain environments