--- a/sqlserver/tests/compose-windows/setup.sql
+++ b/sqlserver/tests/compose-windows/setup.sql
@@ -10,6 +10,7 @@ CREATE USER datadog FOR LOGIN datadog;
 GRANT SELECT on sys.dm_os_performance_counters to datadog;
 GRANT VIEW SERVER STATE to datadog;
 GRANT VIEW ANY DEFINITION to datadog;
+GRANT ALTER SERVER STATE TO datadog;
 
 -- test users
 CREATE LOGIN bob WITH PASSWORD = 'Password12!';