### What does this PR do?
<!-- A brief description of the change being made with this pull request. -->
This is a pull request to change `select plan_handle, statement_start_offset , statement_end_offset` in SQLServer's DBM.
The code added in https://github.com/DataDog/integrations-core/pull/12613 return different results in some environments and an error occur.

ForExample:
- query
```sql
DECLARE @statement_start_offset int  = 14730;
DECLARE @statement_end_offset int = 14838;

SELECT
@statement_start_offset AS original_statement_start_offset,
CONVERT(int, CONVERT(varbinary(4),
    SUBSTRING(
        CONCAT(
        CONVERT(binary(4), @statement_start_offset),
        CONVERT(binary(4), @statement_end_offset) 
        ), 1, 4)
)) AS converted_statement_start_offset,
@statement_end_offset AS original_statement_end_offset,
CONVERT(int, CONVERT(varbinary(4),
    SUBSTRING(
        CONCAT(
        CONVERT(binary(4), @statement_start_offset),
        CONVERT(binary(4), @statement_end_offset)
        ),5, 4)
)) AS converted_statement_end_offset;
SELECT collation_name FROM sys.databases where database_id = DB_ID();
```
- returns
    - Collation : SQL_Latin1_General_CP1_CI_AS
    - ![collation_SQL_Latin1_General_CP1_CI_AS](https://user-images.githubusercontent.com/46398493/235837565-1a7555f4-cfc9-4128-b82c-44e2145f50f6.PNG)
    - Collation : Japanese_CI_AS
![collation_Japanese_CI_AS](https://user-images.githubusercontent.com/46398493/235837574-1c41fbc0-2ff7-443e-8acb-7239cfa56921.PNG)

### Motivation
<!-- What inspired you to submit this pull request? -->
this issue
- https://github.com/DataDog/integrations-core/issues/14006

I can't update datadog agent until fix this error

### Additional Notes
<!-- Anything else we should know when reviewing? -->

### Review checklist (to be filled by reviewers)

- [ ] Feature or bugfix MUST have appropriate tests (unit, integration, e2e)
- [ ] PR title must be written as a CHANGELOG entry [(see why)](https://github.com/DataDog/integrations-core/blob/master/CONTRIBUTING.md#pull-request-title)
- [ ] Files changes must correspond to the primary purpose of the PR as described in the title (small unrelated changes should have their own PR)
- [ ] PR must have `changelog/` and `integration/` labels attached
- [ ] If the PR doesn't need to be tested during QA, please add a `qa/skip-qa` label.