@nenadnoveljic 2023-05-31
It seems like the change broke some tests:

```
=================================== FAILURES ===================================
________ test_statement_metadata[metadata1-expected_metadata_payload1] _________
tests/test_statements.py:482: in test_statement_metadata
    assert sample['db']['metadata']['comments'] == expected_metadata_payload['comments']
E   AssertionError: assert ['-- Test comment'] == None
------------------------------ Captured log call -------------------------------
WARNING  datadog_checks.base.checks.base.sqlserver:sqlserver.py:189 Autodiscovery is disabled, autodiscovery_include and autodiscovery_exclude will be ignored
WARNING  datadog_checks.base.checks.base.sqlserver:core.py:247 QueryManager initialized with no query
______________________ test_statement_basic_metrics_query ______________________
tests/test_statements.py:678: in test_statement_basic_metrics_query
    assert matching, "the test query should be visible in the query stats"
E   AssertionError: the test query should be visible in the query stats
E   assert []
- generated xml file: /home/runner/work/integrations-core/integrations-core/sqlserver/.junit/test-unit-py3.8-linux-2017-single$TOX_ENV_NAME.xml -
=========================== short test summary info ============================
FAILED tests/test_statements.py::test_statement_metadata[metadata1-expected_metadata_payload1] - AssertionError: assert ['-- Test comment'] == None
FAILED tests/test_statements.py::test_statement_basic_metrics_query - AssertionError: the test query should be visible in the query stats
assert []
============ 2 failed, 192 passed, 20 skipped in 322.33s (0:05:22) =============
```