@it-ito 2023-06-01
@nenadnoveljic 
When there is a cache of queries with the same `query_hash` and `query_plan_hash`, [this max(plan_handle_and_offsets)](https://github.com/DataDog/integrations-core/blob/adf5d2a523ca41945747eb858c092a0d39523b01/sqlserver/datadog_checks/sqlserver/statements.py#L68)  may be different from before.
What do you think of the fix commit (ead85416daec947c7262cf16b78c8f6f3aa12e82) that would keep this from changing?

If the possibility of changing the results is acceptable, the test case can be modified.
`tests/test_statements.py::test_statement_basic_metrics_query` is affected by other tests, so if `DBCC FREEPROCCACHE` is available, this case would be a success.
But I don't know how to fix `tests/test_statements.py::test_statement_metadata[metadata1-expected_metadata_payload1]` .