@nenadnoveljic 2023-06-02
Thank you for the changes!

I compared the performance to the old query. The new query performs better on my test database - lots of logical reads from the Worktable are eliminated:

Old query run 1:
```
Started executing query at Line 5
SQL Server parse and compile time: 
   CPU time = 44 ms, elapsed time = 44 ms.
(26 rows affected)
Table 'Worktable'. Scan count 735, logical reads 2677, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'Workfile'. Scan count 0, logical reads 0, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'sysobjvalues'. Scan count 162, logical reads 380, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'syssingleobjrefs'. Scan count 81, logical reads 162, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'sysdbreg'. Scan count 93, logical reads 186, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
```

Old query run 2:

```
Started executing query at Line 5
SQL Server parse and compile time: 
   CPU time = 0 ms, elapsed time = 0 ms.
(21 rows affected)
Table 'Worktable'. Scan count 736, logical reads 2683, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'Workfile'. Scan count 0, logical reads 0, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'sysobjvalues'. Scan count 162, logical reads 380, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'syssingleobjrefs'. Scan count 81, logical reads 162, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'sysdbreg'. Scan count 93, logical reads 186, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
```

New query run 1:

```
Started executing query at Line 55
SQL Server parse and compile time: 
   CPU time = 75 ms, elapsed time = 75 ms.
(44 rows affected)
Table 'Worktable'. Scan count 0, logical reads 0, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'Workfile'. Scan count 0, logical reads 0, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'sysobjvalues'. Scan count 160, logical reads 376, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'syssingleobjrefs'. Scan count 80, logical reads 160, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'sysdbreg'. Scan count 85, logical reads 170, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
```

New query run 2:
```
Started executing query at Line 54
SQL Server parse and compile time: 
   CPU time = 0 ms, elapsed time = 0 ms.
(24 rows affected)
Table 'Worktable'. Scan count 0, logical reads 0, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'Workfile'. Scan count 0, logical reads 0, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'sysobjvalues'. Scan count 160, logical reads 376, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'syssingleobjrefs'. Scan count 80, logical reads 160, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.
Table 'sysdbreg'. Scan count 85, logical reads 170, physical reads 0, read-ahead reads 0, lob logical reads 0, lob physical reads 0, lob read-ahead reads 0.

 SQL Server Execution Times:
   CPU time = 16 ms,  elapsed time = 15 ms.
Total execution time: 00:00:00.276
```

Would it be too much to ask that you perform the same test on one of your databases, preferably one with lots of statements in cache? This SQL is one of the heaviest in the agent, so just wanted to make sure that the change doesn't have adversary effects. 

The test consists of turning the diagnostics on in SQL Server Management Studio:

```
set statistics IO ON
set statistics time on
```

And then run the following two statements two times each.

Old query:

```
with qstats as (
    select query_hash, query_plan_hash, last_execution_time, last_elapsed_time,
            CONCAT(
                CONVERT(binary(64), plan_handle),
                CONVERT(binary(4), statement_start_offset),
                CONVERT(binary(4), statement_end_offset)) as plan_handle_and_offsets_order,
            CONCAT(
                CONVERT(VARCHAR(64), CONVERT(binary(64), plan_handle), 1),
                CONVERT(VARCHAR(10), CONVERT(varbinary(4), statement_start_offset), 1),
                CONVERT(VARCHAR(10), CONVERT(varbinary(4), statement_end_offset), 1)
                ) as plan_handle_and_offsets_varchar,
           (select value from sys.dm_exec_plan_attributes(plan_handle) where attribute = 'dbid') as dbid,
            execution_count, total_worker_time as total_worker_time, total_physical_reads, total_logical_writes, total_logical_reads, total_clr_time, total_elapsed_time, total_rows, total_dop, total_grant_kb, total_used_grant_kb, total_ideal_grant_kb, total_reserved_threads, total_used_threads, total_columnstore_segment_reads, total_columnstore_segment_skips, total_spills
    from sys.dm_exec_query_stats
),
qstats_aggr_order as (
    select FIRST_VALUE(plan_handle_and_offsets_varchar)
        over (partition by query_hash, query_plan_hash, S.dbid
            order by plan_handle_and_offsets_order desc) as plan_handle_and_offsets,
        S.*
    from qstats as S
),
qstats_aggr as (
    select query_hash, query_plan_hash, CAST(S.dbid as int) as dbid,
       D.name as database_name, max(plan_handle_and_offsets) as plan_handle_and_offsets,
       max(last_execution_time) as last_execution_time,
       max(last_elapsed_time) as last_elapsed_time,
       sum(execution_count) as execution_count, sum(total_worker_time) as total_worker_time, sum(total_physical_reads) as total_physical_reads, sum(total_logical_writes) as total_logical_writes, sum(total_logical_reads) as total_logical_reads, sum(total_clr_time) as total_clr_time, sum(total_elapsed_time) as total_elapsed_time, sum(total_rows) as total_rows, sum(total_dop) as total_dop, sum(total_grant_kb) as total_grant_kb, sum(total_used_grant_kb) as total_used_grant_kb, sum(total_ideal_grant_kb) as total_ideal_grant_kb, sum(total_reserved_threads) as total_reserved_threads, sum(total_used_threads) as total_used_threads, sum(total_columnstore_segment_reads) as total_columnstore_segment_reads, sum(total_columnstore_segment_skips) as total_columnstore_segment_skips, sum(total_spills) as total_spills
    from qstats_aggr_order S
    left join sys.databases D on S.dbid = D.database_id
    group by query_hash, query_plan_hash, S.dbid, D.name
),
qstats_aggr_split as (select TOP 10000
    convert(varbinary(64), convert(binary(64), substring(plan_handle_and_offsets, 1, 64), 1)) as plan_handle,
    convert(int, convert(varbinary(10), substring(plan_handle_and_offsets, 64+1, 10), 1)) as statement_start_offset,
    convert(int, convert(varbinary(10), substring(plan_handle_and_offsets, 64+11, 10), 1)) as statement_end_offset,
    * from qstats_aggr
    where DATEADD(ms, last_elapsed_time / 1000, last_execution_time) > dateadd(second, -60, getdate())
)
select
    SUBSTRING(text, (statement_start_offset / 2) + 1,
    ((CASE statement_end_offset
        WHEN -1 THEN DATALENGTH(text)
        ELSE statement_end_offset END
            - statement_start_offset) / 2) + 1) AS statement_text,
    qt.text,
    encrypted as is_encrypted,
    * from qstats_aggr_split
    cross apply sys.dm_exec_sql_text(plan_handle) qt;
```

New query:

```
with qstats as (
    select query_hash, query_plan_hash, last_execution_time, last_elapsed_time,
            CONCAT(
                CONVERT(binary(64), plan_handle),
                CONVERT(binary(4), statement_start_offset),
                CONVERT(binary(4), statement_end_offset)) as plan_handle_and_offsets,
           (select value from sys.dm_exec_plan_attributes(plan_handle) where attribute = 'dbid') as dbid,
                       execution_count, total_worker_time as total_worker_time, total_physical_reads, total_logical_writes, total_logical_reads, total_clr_time, total_elapsed_time, total_rows, total_dop, total_grant_kb, total_used_grant_kb, total_ideal_grant_kb, total_reserved_threads, total_used_threads, total_columnstore_segment_reads, total_columnstore_segment_skips, total_spills
    from sys.dm_exec_query_stats
),
qstats_aggr as (
    select query_hash, query_plan_hash, CAST(S.dbid as int) as dbid,
       D.name as database_name, max(plan_handle_and_offsets) as plan_handle_and_offsets,
       max(last_execution_time) as last_execution_time,
       max(last_elapsed_time) as last_elapsed_time,
    sum(execution_count) as execution_count, sum(total_worker_time) as total_worker_time, sum(total_physical_reads) as total_physical_reads, sum(total_logical_writes) as total_logical_writes, sum(total_logical_reads) as total_logical_reads, sum(total_clr_time) as total_clr_time, sum(total_elapsed_time) as total_elapsed_time, sum(total_rows) as total_rows, sum(total_dop) as total_dop, sum(total_grant_kb) as total_grant_kb, sum(total_used_grant_kb) as total_used_grant_kb, sum(total_ideal_grant_kb) as total_ideal_grant_kb, sum(total_reserved_threads) as total_reserved_threads, sum(total_used_threads) as total_used_threads, sum(total_columnstore_segment_reads) as total_columnstore_segment_reads, sum(total_columnstore_segment_skips) as total_columnstore_segment_skips, sum(total_spills) as total_spills
    from qstats S
    left join sys.databases D on S.dbid = D.database_id
    group by query_hash, query_plan_hash, S.dbid, D.name
),
qstats_aggr_split as (select TOP 10000
    convert(varbinary(64), substring(plan_handle_and_offsets, 1, 64)) as plan_handle,
    convert(int, convert(varbinary(4), substring(plan_handle_and_offsets, 64+1, 4))) as statement_start_offset,
    convert(int, convert(varbinary(4), substring(plan_handle_and_offsets, 64+6, 4))) as statement_end_offset,
    * from qstats_aggr
    where DATEADD(ms, last_elapsed_time / 1000, last_execution_time) > dateadd(second, -60, getdate())
)
select
    SUBSTRING(text, (statement_start_offset / 2) + 1,
    ((CASE statement_end_offset
        WHEN -1 THEN DATALENGTH(text)
        ELSE statement_end_offset END
            - statement_start_offset) / 2) + 1) AS statement_text,
    qt.text,
    encrypted as is_encrypted,
    * from qstats_aggr_split
    cross apply sys.dm_exec_sql_text(plan_handle) qt
```

If it's possible for you to do this test, please could you post the results here? They are in the tab "Messages".

Further, there's another similar query - stored in the variable `STATEMENT_METRICS_QUERY_NO_AGGREGATES`, which probably suffers the same issue. I don't know if you are affected by this. If not, we can do the changes.